backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
openai-whisper==20231117
torch==2.3.1
TTS==0.22.0
deep-translator==1.11.4
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.7.4
numpy==1.26.4
soundfile==0.12.1